Basic_Software_Programming/python/Projects/crud_flask_apirest/requirements.txt
flask
flask_sqlalchemy
mysql-conector-python
mysqlcient